Correct config script

+++ scripts/config-setup.js
@@ -5,7 +5,10 @@ const fileExists = async path => !!(await fs.stat(path).catch(e => false));
 const target = '.temp_custom_deps';
 const main = async () => {
 
-    await fs.rm(target, { recursive: true, force: true });
+    if (await fileExists(configLocation)){
+        await fs.rm(target, { recursive: true, force: true });
+    }
+   
     const configLocation = process.env.CONFIG_LOCAION ?? './mimir-config';
     if (await fileExists(configLocation)) {
         console.log(`Copying ${configLocation} to ${target}}`)
